+++ pydantic_forms/validators/components/list_of_two.py
@@ -1,0 +1,22 @@
+# Copyright 2019-2023 SURF.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from typing import TypeVar
+
+from pydantic_forms.validators.components.unique_constrained_list import UniqueConstrainedList
+
+T = TypeVar("T")  # pragma: no mutate
+
+
+class ListOfTwo(UniqueConstrainedList[T]):
+    min_items = 2
+    max_items = 2
