+++ tests/unit_tests/test_read_only_field.py
@@ -3,21 +3,23 @@ import pytest
 from pydantic_forms.core import FormPage, ReadOnlyField
 
 
-@pytest.mark.xfail(reason="fix flakey test")
+# @pytest.mark.xfail(reason="fix flakey test")
 def test_read_only_field_schema():
     class Form(FormPage):
         read_only: int = ReadOnlyField(1, const=False)
 
-    assert Form.schema() == {
+    expected = {
         "title": "unknown",
         "type": "object",
         "properties": {
             "read_only": {
+                "default": 1,
                 "title": "Read Only",
-                "const": 1,
                 "uniforms": {"disabled": True, "value": 1},
                 "type": "integer",
             }
         },
         "additionalProperties": False,
     }
+
+    assert Form.model_json_schema() == expected
